Improved comments; added readme; made changes.
This code was poorly commented, so I added a lot of comments
clarifying what's going on.

Additionally, there was no readme. I added detailed instructions
on how to use both the default and advanced modes of this program.

Lastly, I updated the XML update utility to be more consistent,
and to work under more advanced scenarios.

I actually didn't have to update the comparator; I apparently already
did that.